Sets the install location using PREFIX
This replaces the DESTDIR variable with PREFIX and importantly,
removes "usr/" that was prepended to DESTDIR. This allows installation
to places like /usr/local or $HOME/.local

No default PREFIX is assumed, if not provided the installation will
exit with error.

The uninstall target no longer assumes the installation is at /usr
and instead requires the PREFIX to be set.

The installed desktop file is modified to use the full path to the script.

Warnings about not being root have been removed.